Drop Rig functionality complete
Created a file and game object that stores the information about the planets.
This objects purpose is to provide a single place to put the information in that will help set up the planet attributes like gravity when the planet is loaded. Its currently setting up the gravity and atmosphere parameters for sound.
This will also provide the means for the information on the wrist watch UI to get its informaiton.
For example, this object will set the gravity and if the sounds will play for the object as they collide, the button and the drop rig sounds. The particles will use this files hasAtmos variable to calculate if the particles created when an object hits the ground should drift in the air or fall strait down, but they don’t do this yet.
This means that the Gravity, Air drag, and sound propagations is all possible to switch now with a single toggle box for isMoon, isEarth or is Mars. You will just need to add this prefab to your planet’s scenes, and it will set it all up at runtime.
I have added the ability for object to play sound effect when they collide with things if the planet has an atmosphere. The same for the drop rig height change and the drop wings tigering. They all make sound.
I have added the ability for objects to have a particle effect when they collide with the ground, but I haven’t created the particle emitter as of yet. There is just a place holder piece of code for now that displays something to the console and records the position the particles should be spawned at.
I have also created the placeholder code for the objects falling and they now keep a record on them selves about when they began to drop and how long until they hit the ground. They are all ready for the LCD screen UI to be created. This same information is ready to be put into the wrist watch UI if we decide to put it there.
I also prevented massive numbers of objects from being spawned at once. This can be turned off if the user wants.
I have commented my code so that its easier to understand.